wrap bort in server component
